pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kronos-bot"
version = "1.0.0"
description = "Kronos AI trading bot for Hyperliquid perpetuals"
requires-python = ">=3.11"
dependencies = [
    "hyperliquid-python-sdk>=0.6.0",
    "pandas>=2.0.0",
    "pandas-ta>=0.3.14b0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "huggingface-hub>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "ANN",   # flake8-annotations
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
filterwarnings = ["ignore::DeprecationWarning"]